+++ polars/polars-lazy/polars-plan/src/dsl/function_expr/datetime.rs
@@ -206,32 +206,17 @@ pub(super) fn timestamp(s: &Series, tu: TimeUnit) -> PolarsResult<Series> {
 }
 
 pub(super) fn truncate(s: &Series, options: &TruncateOptions) -> PolarsResult<Series> {
-    let every = Duration::parse(&options.every);
-    let offset = Duration::parse(&options.offset);
     let mut out = match s.dtype() {
         DataType::Datetime(_, tz) => match tz {
             #[cfg(feature = "timezones")]
             Some(tz) => s
                 .datetime()
                 .unwrap()
-                .truncate(
-                    every,
-                    offset,
-                    tz.parse::<Tz>().ok().as_ref(),
-                    options.use_earliest,
-                )?
+                .truncate(options, tz.parse::<Tz>().ok().as_ref())?
                 .into_series(),
-            _ => s
-                .datetime()
-                .unwrap()
-                .truncate(every, offset, None, None)?
-                .into_series(),
+            _ => s.datetime().unwrap().truncate(options, None)?.into_series(),
         },
-        DataType::Date => s
-            .date()
-            .unwrap()
-            .truncate(every, offset, None, None)?
-            .into_series(),
+        DataType::Date => s.date().unwrap().truncate(options, None)?.into_series(),
         dt => polars_bail!(opq = round, got = dt, expected = "date/datetime"),
     };
     out.set_sorted_flag(s.is_sorted_flag());
